refactor(admin_ops): observability 335\u2192{99,129,114} splits
