+++ src/bin/email_api_dir/admin_ops/diag_handlers/mod.rs
@@ -3,11 +3,15 @@
 pub mod deliverability;
 pub mod deliverability_builders;
 pub mod observability;
+pub mod observability_alerts;
 pub mod observability_collectors;
+pub mod observability_stats;
 pub mod security;
 
 pub use deliverability::*;
 pub use deliverability_builders::*;
 pub use observability::*;
+pub use observability_alerts::*;
 pub use observability_collectors::*;
+pub use observability_stats::*;
 pub use security::*;
