+++ src/bin/email_api_dir/admin_ops/diag_handlers/observability.rs
@@ -1,244 +1,7 @@
 #![allow(unused_imports, dead_code)]
 use super::super::*;
-
-/// Résumé des compteurs SMTP par statut sur la fenêtre.
-struct SmtpStatusStats {
-    total: u64,
-    delivered: u64,
-    bounced: u64,
-    failed: u64,
-    deferred: u64,
-    by_status: serde_json::Map<String, serde_json::Value>,
-    p95: Option<u64>,
-}
-
-async fn collect_smtp_stats(
-    mongo: &Arc<mongodb::Client>,
-    since: &str,
-    base_filter: &bson::Document,
-) -> SmtpStatusStats {
-    let total = storage::count_events(mongo, base_filter.clone()).await;
-    let by_status_docs = storage::aggregate(
-        mongo,
-        vec![
-            doc! { "$match": base_filter.clone() },
-            doc! { "$group": { "_id": "$status", "count": { "$sum": 1 }, "avg_ms": { "$avg": "$total_ms" } } },
-        ],
-    )
-    .await;
-    let mut by_status = serde_json::Map::new();
-    let (mut delivered, mut bounced, mut failed, mut deferred) = (0u64, 0u64, 0u64, 0u64);
-    for doc in &by_status_docs {
-        let status = doc.get_str("_id").unwrap_or("unknown").to_string();
-        let count = doc.get_i64("count").unwrap_or(0) as u64;
-        match status.as_str() {
-            "delivered" => delivered = count,
-            "bounced" => bounced = count,
-            "failed" => failed = count,
-            "deferred" => deferred = count,
-            _ => {}
-        }
-        by_status.insert(status, serde_json::json!(count));
-    }
-    let p95 = storage::p95_total_ms(mongo, base_filter.clone(), 1000).await;
-    let _ = since;
-    SmtpStatusStats { total, delivered, bounced, failed, deferred, by_status, p95 }
-}
-
-struct QueueStats {
-    depth: u64,
-    oldest_age_seconds: Option<u64>,
-}
-
-async fn collect_queue_stats(mongo: &Arc<mongodb::Client>) -> QueueStats {
-    let db = env::var("MONGODB_DATABASE").unwrap_or_else(|_| "mailserver".to_string());
-    let queue_coll = mongo
-        .database(&db)
-        .collection::<bson::Document>(SEND_QUEUE_COLL);
-    let pending_queue_filter = doc! { "status": { "$in": ["pending", "scheduled", "sending"] } };
-    let depth = queue_coll
-        .count_documents(pending_queue_filter.clone())
-        .await
-        .unwrap_or(0);
-    let oldest_pending = queue_coll
-        .find_one(pending_queue_filter)
-        .sort(doc! { "created_at": 1 })
-        .await
-        .ok()
-        .flatten();
-    let oldest_age_seconds = oldest_pending
-        .as_ref()
-        .and_then(|d| d.get_datetime("created_at").ok())
-        .map(|dt| (Utc::now().timestamp_millis() - dt.timestamp_millis()).max(0) as u64 / 1000);
-    QueueStats { depth, oldest_age_seconds }
-}
-
-struct ThroughputStats {
-    incoming: u64,
-    outgoing: u64,
-    smtp_4xx: u64,
-    smtp_5xx: u64,
-    dns_issue_events: u64,
-}
-
-async fn collect_throughput_stats(
-    mongo: &Arc<mongodb::Client>,
-    since: &str,
-) -> ThroughputStats {
-    let incoming = storage::count_events(
-        mongo,
-        doc! { "ts": { "$gte": since }, "event_type": { "$in": ["accepted", "received"] } },
-    )
-    .await;
-    let outgoing = storage::count_events(
-        mongo,
-        doc! { "ts": { "$gte": since }, "status": { "$in": ["delivered", "bounced", "failed", "deferred"] } },
-    )
-    .await;
-    let smtp_4xx = storage::count_events(
-        mongo,
-        doc! { "ts": { "$gte": since }, "smtp_code": { "$gte": 400, "$lt": 500 } },
-    )
-    .await;
-    let smtp_5xx = storage::count_events(
-        mongo,
-        doc! { "ts": { "$gte": since }, "smtp_code": { "$gte": 500, "$lt": 600 } },
-    )
-    .await;
-    let dns_issue_events = storage::count_events(
-        mongo,
-        doc! { "ts": { "$gte": since }, "event_type": "dns_lookup", "status": { "$in": ["failed", "deferred"] } },
-    )
-    .await;
-    ThroughputStats { incoming, outgoing, smtp_4xx, smtp_5xx, dns_issue_events }
-}
-
-struct AlertsSnapshot {
-    monitoring: Vec<monitoring::alerts::ActiveAlert>,
-    security: Vec<simple_smtp_server::security::SecurityAlert>,
-    queue_growth: usize,
-    auth_failures: usize,
-    anomalies: usize,
-}
-
-async fn collect_alerts_snapshot(
-    mongo: &Arc<mongodb::Client>,
-    window: &str,
-) -> AlertsSnapshot {
-    use simple_smtp_server::security::audit;
-    let monitoring = monitoring::alerts::evaluate_alerts(
-        mongo,
-        parse_window(window).num_minutes(),
-        &AlertConfig::default(),
-    )
-    .await;
-    let security = audit::query_active_alerts(mongo, 300).await;
-    let queue_growth = monitoring
-        .iter()
-        .filter(|a| a.kind.contains("queue") || a.message.to_ascii_lowercase().contains("queue"))
-        .count();
-    let auth_failures = security
-        .iter()
-        .filter(|a| {
-            let n = a.rule_name.to_ascii_lowercase();
-            n.contains("auth") || n.contains("brute") || n.contains("login")
-        })
-        .count();
-    let anomalies = security
-        .iter()
-        .filter(|a| {
-            let n = a.rule_name.to_ascii_lowercase();
-            n.contains("volume")
-                || n.contains("spike")
-                || n.contains("anormal")
-                || n.contains("anomaly")
-        })
-        .count();
-    AlertsSnapshot { monitoring, security, queue_growth, auth_failures, anomalies }
-}
-
-async fn collect_suspicious_logins(
-    mongo: &Arc<mongodb::Client>,
-    since: &str,
-) -> Vec<serde_json::Value> {
-    let db = env::var("MONGODB_DATABASE").unwrap_or_else(|_| "mailserver".to_string());
-    let auth_events_coll = mongo
-        .database(&db)
-        .collection::<bson::Document>("auth_events");
-    let docs = match auth_events_coll
-        .aggregate(vec![
-            doc! { "$match": { "ts": { "$gte": since }, "success": false } },
-            doc! { "$group": { "_id": "$ip", "attempts": { "$sum": 1 } } },
-            doc! { "$sort": { "attempts": -1 } },
-            doc! { "$limit": 10 },
-        ])
-        .await
-    {
-        Ok(cursor) => cursor.try_collect::<Vec<_>>().await.unwrap_or_default(),
-        Err(_) => Vec::new(),
-    };
-    docs.into_iter()
-        .map(|d| {
-            serde_json::json!({
-                "ip": d.get_str("_id").unwrap_or("unknown"),
-                "attempts": d.get_i64("attempts").unwrap_or(0)
-            })
-        })
-        .collect()
-}
-
-async fn collect_per_domain(
-    mongo: &Arc<mongodb::Client>,
-    base_filter: &bson::Document,
-) -> Vec<serde_json::Value> {
-    let docs = storage::aggregate(
-        mongo,
-        vec![
-            doc! { "$match": base_filter.clone() },
-            doc! { "$project": {
-                "recipient_domain": { "$arrayElemAt": [ { "$split": ["$to", "@"] }, 1 ] },
-                "status": "$status"
-            }},
-            doc! { "$group": {
-                "_id": "$recipient_domain",
-                "count": { "$sum": 1 },
-                "delivered": { "$sum": { "$cond": { "if": { "$eq": ["$status", "delivered"] }, "then": 1, "else": 0 } } },
-                "bounced": { "$sum": { "$cond": { "if": { "$eq": ["$status", "bounced"] }, "then": 1, "else": 0 } } }
-            }},
-            doc! { "$sort": { "count": -1 } },
-            doc! { "$limit": 20 }
-        ],
-    )
-    .await;
-    docs.into_iter()
-        .map(|d| {
-            serde_json::json!({
-                "domain": d.get_str("_id").unwrap_or("unknown"),
-                "count": d.get_i64("count").unwrap_or(0),
-                "delivered": d.get_i64("delivered").unwrap_or(0),
-                "bounced": d.get_i64("bounced").unwrap_or(0)
-            })
-        })
-        .collect()
-}
-
-fn rbl_sources() -> Vec<String> {
-    env::var("RBL_CHECK_HOSTS")
-        .unwrap_or_else(|_| "zen.spamhaus.org,bl.spamcop.net".to_string())
-        .split(',')
-        .map(|s| s.trim().to_string())
-        .filter(|s| !s.is_empty())
-        .collect()
-}
-
-fn rbl_listed_by() -> Vec<String> {
-    env::var("RBL_LISTED_BY")
-        .unwrap_or_default()
-        .split(',')
-        .map(|s| s.trim().to_string())
-        .filter(|s| !s.is_empty())
-        .collect()
-}
+use super::observability_alerts::*;
+use super::observability_stats::*;
 
 pub(crate) async fn api_admin_observability_overview(
     query: web::Query<AdminWindowQuery>,
